+++ src/Editor/Exceptions/EditorDocumentNotOwned.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atatusoft\Ppphp\Editor\Exceptions;
+
+final class EditorDocumentNotOwned extends \InvalidArgumentException {}